refactor(web): rename /annual to /cars/annual
