+++ apps/web/src/app/(main)/(explore)/cars/annual/components/car-population-chart.tsx
@@ -7,7 +7,7 @@ import {
   ChartTooltip,
   ChartTooltipContent,
 } from "@sgcarstrends/ui/components/chart";
-import { useEffectiveYear } from "@web/app/(main)/(explore)/annual/hooks/use-effective-year";
+import { useEffectiveYear } from "@web/app/(main)/(explore)/cars/annual/hooks/use-effective-year";
 import Typography from "@web/components/typography";
 import {
   CARD_PADDING,
